requirements.txt
pandas==2.1.0
streamlit==1.28.0
plotly==5.17.0
numpy==1.24.3
spotipy==2.23.0
python-dotenv==1.0.0